move container components to containers folder
