+++ src/containers/BooksList.js
@@ -1,7 +1,7 @@
 import { nanoid } from 'nanoid';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import Book from './Book';
+import Book from '../components/Book';
 
 function BooksList(props) {
   const { books } = props;
